# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/courses": {
      "get": {
        "description": "Público (token opcional). RN23–RN27. scope=catalog mostra só publicados; scope=mine é do professor; scope=all é do admin.",
        "operationId": "Courses_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 12,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Busca parcial em título e resumo, sem diferenciar maiúsculas/minúsculas e acentos",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Slug da categoria",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "BEGINNER",
                "INTERMEDIATE",
                "ADVANCED"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Só tem efeito com scope=mine ou scope=all",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "PUBLISHED",
                "ARCHIVED"
              ]
            }
          },
          {
            "name": "teacherId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "newest",
              "type": "string",
              "enum": [
                "newest",
                "oldest",
                "title",
                "popular"
              ]
            }
          },
          {
            "name": "scope",
            "required": false,
            "in": "query",
            "description": "catalog = publicados; mine = do professor logado; all = todos (admin)",
            "schema": {
              "default": "catalog",
              "type": "string",
              "enum": [
                "catalog",
                "mine",
                "all"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página de resultados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CourseSummaryResponse"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `INVALID_SCOPE`. scope=mine sem ser professor ou scope=all sem ser admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "INVALID_SCOPE": {
                    "summary": "INVALID_SCOPE",
                    "value": {
                      "statusCode": 403,
                      "code": "INVALID_SCOPE",
                      "message": "scope=all é exclusivo de administradores",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Lista cursos com busca, filtros, ordenação e paginação",
        "tags": [
          "Courses"
        ]
      },
      "post": {
        "description": "RN15–RN16. O slug é gerado do título e é único.",
        "operationId": "Courses_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada — `INVALID_CATEGORY`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "INVALID_CATEGORY": {
                    "summary": "INVALID_CATEGORY",
                    "value": {
                      "statusCode": 422,
                      "code": "INVALID_CATEGORY",
                      "message": "Categoria informada não existe",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cria curso em rascunho",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/courses/{idOrSlug}": {
      "get": {
        "description": "Público (token opcional). RN28–RN29. Rascunho só é visível ao dono e ao admin (os demais recebem 404). Com token, inclui o contexto do usuário em `viewer`.",
        "operationId": "Courses_detail",
        "parameters": [
          {
            "name": "idOrSlug",
            "required": true,
            "in": "path",
            "schema": {
              "example": "playwright-com-typescript-do-zero",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado — `NOT_FOUND`. Curso inexistente ou rascunho de outra pessoa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "statusCode": 404,
                      "code": "NOT_FOUND",
                      "message": "Curso não encontrado(a)",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detalhe do curso por ID ou slug",
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/api/courses/{id}/publish": {
      "post": {
        "description": "Apenas o professor dono do curso ou um administrador. RN17–RN18. Exige ao menos 1 aula.",
        "operationId": "Courses_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado — `NOT_FOUND`. Curso inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "statusCode": 404,
                      "code": "NOT_FOUND",
                      "message": "Curso não encontrado(a)",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada — `INVALID_STATUS_TRANSITION`, `COURSE_WITHOUT_LESSONS`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "INVALID_STATUS_TRANSITION": {
                    "summary": "INVALID_STATUS_TRANSITION",
                    "value": {
                      "statusCode": 422,
                      "code": "INVALID_STATUS_TRANSITION",
                      "message": "Apenas cursos em rascunho podem ser publicados",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  },
                  "COURSE_WITHOUT_LESSONS": {
                    "summary": "COURSE_WITHOUT_LESSONS",
                    "value": {
                      "statusCode": 422,
                      "code": "COURSE_WITHOUT_LESSONS",
                      "message": "Adicione ao menos uma aula antes de publicar",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "DRAFT → PUBLISHED",
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/courses/{id}/archive": {
      "post": {
        "description": "Apenas o professor dono do curso ou um administrador. RN17. Encerra novas matrículas; o curso vira somente leitura.",
        "operationId": "Courses_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseDetailResponse"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos — `VALIDATION_ERROR`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "VALIDATION_ERROR": {
                    "summary": "VALIDATION_ERROR",
                    "value": {
                      "statusCode": 400,
                      "code": "VALIDATION_ERROR",
                      "message": "Dados inválidos",
                      "details": [
                        {
                          "field": "campo",
                          "message": "mensagem de validação"
                        }
                      ],
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado — `UNAUTHORIZED`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "UNAUTHORIZED": {
                    "summary": "UNAUTHORIZED",
                    "value": {
                      "statusCode": 401,
                      "code": "UNAUTHORIZED",
                      "message": "Autenticação necessária",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão — `FORBIDDEN`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "FORBIDDEN": {
                    "summary": "FORBIDDEN",
                    "value": {
                      "statusCode": 403,
                      "code": "FORBIDDEN",
                      "message": "Seu perfil não tem permissão para esta ação",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado — `NOT_FOUND`. Curso inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "NOT_FOUND": {
                    "summary": "NOT_FOUND",
                    "value": {
                      "statusCode": 404,
                      "code": "NOT_FOUND",
                      "message": "Curso não encontrado(a)",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada — `INVALID_STATUS_TRANSITION`",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                },
                "examples": {
                  "INVALID_STATUS_TRANSITION": {
                    "summary": "INVALID_STATUS_TRANSITION",
                    "value": {
                      "statusCode": 422,
                      "code": "INVALID_STATUS_TRANSITION",
                      "message": "Apenas cursos em rascunho podem ser publicados",
                      "path": "/api/...",
                      "timestamp": "2026-09-28T12:00:00.000Z",
                      "requestId": "8f0c2a4e-1b7d-4c1e-9a55-2f6c0e9d1a3b"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "PUBLISHED → ARCHIVED",
        "tags": [
          "Courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "EduTrack API",
    "description": "API da plataforma EduTrack: cursos, aulas, matrículas, progresso, notas e certificados.\n\n### Autenticação\n1. `POST /api/auth/login` com `aluno@edutrack.dev` / `Senha@123` (ou outro usuário do seed).\n2. Copie o `accessToken` e clique em **Authorize**.\n\nRotas marcadas como **públicas** aceitam token opcional (com token, a resposta traz o contexto do usuário).\n\n### Erros\nSempre no formato `ErrorResponseDto`: `{ statusCode, code, message, details?, path, timestamp, requestId }`.\nFaça asserções pelo **`code`** (estável); `message` é texto para humanos.\n\n| Status | Significado |\n|---|---|\n| 400 | Dados inválidos (`details` lista os campos) |\n| 401 | Não autenticado / credenciais inválidas |\n| 403 | Autenticado, mas sem permissão |\n| 404 | Não existe ou não é visível para você |\n| 409 | Conflito de estado ou unicidade |\n| 422 | Regra de negócio violada |\n| 429 | Limite de tentativas de login |\n\nToda resposta traz o header `X-Request-Id`, que também aparece no log do servidor.\n\nAs regras citadas como RNxx estão em `docs/regras-de-negocio.md`.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "accessToken obtido no login"
      }
    },
    "schemas": {
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 12
          },
          "total": {
            "type": "number",
            "example": 42
          },
          "totalPages": {
            "type": "number",
            "example": 4
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ]
      },
...
      "CourseSummaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Playwright com TypeScript do Zero"
          },
          "slug": {
            "type": "string",
            "example": "playwright-com-typescript-do-zero"
          },
          "summary": {
            "type": "string",
            "example": "Automação de testes web moderna."
          },
          "level": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseLevel"
              }
            ]
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CourseStatus"
              }
            ]
          },
          "workloadHours": {
            "type": "number",
            "example": 40
          },
          "capacity": {
            "type": "number",
            "example": 50
          },
          "enrolledCount": {
            "type": "number",
            "example": 2,
            "description": "Matrículas ativas"
          },
          "seatsLeft": {
            "type": "number",
            "example": 48
          },
          "lessonsCount": {
            "type": "number",
            "example": 6
          },
          "totalMinutes": {
            "type": "number",
            "example": 127,
            "description": "Soma da duração das aulas, em minutos"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryRef"
          },
          "teacher": {
            "$ref": "#/components/schemas/PersonRef"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "summary",
          "level",
          "status",
          "workloadHours",
          "capacity",
          "enrolledCount",
          "seatsLeft",
          "lessonsCount",
          "totalMinutes",
          "category",
          "teacher",
          "publishedAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  }
}